implementa sistema de roles para todas as rotas

+++ src/main/java/com/pedro/petshop/controllers/AddressController.java
@@ -5,6 +5,8 @@ import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -14,11 +16,14 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.pedro.petshop.configs.CustomAuthentication;
 import com.pedro.petshop.configs.RolesAllowed;
 import com.pedro.petshop.dtos.AddressDTO;
 import com.pedro.petshop.entities.Address;
+import com.pedro.petshop.enums.Role;
 import com.pedro.petshop.mappers.AddressMapper;
 import com.pedro.petshop.services.AddressService;
+import com.pedro.petshop.services.ClientService;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -31,10 +36,13 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 public class AddressController {
 
         private final AddressService addressService;
+        private final ClientService clientService;
         private final AddressMapper addressMapper;
 
-        public AddressController(AddressService addressService, AddressMapper addressMapper) {
+        public AddressController(AddressService addressService, AddressMapper addressMapper,
+                        ClientService clientService) {
                 this.addressService = addressService;
+                this.clientService = clientService;
                 this.addressMapper = addressMapper;
         }
 
@@ -47,8 +55,22 @@ public class AddressController {
         })
         @RolesAllowed({ "ADMIN" })
         @PostMapping
-        public AddressDTO createAddress(@RequestBody AddressDTO address) {
-                return addressMapper.toDto(addressService.create(addressMapper.toEntity(address)));
+        public ResponseEntity<AddressDTO> createAddress(@RequestBody AddressDTO address) {
+                Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+                if (authentication instanceof CustomAuthentication) {
+                        CustomAuthentication customAuth = (CustomAuthentication) authentication;
+                        String role = customAuth.getRole();
+                        String cpf = customAuth.getCpf();
+
+                        if (role.equals(Role.CLIENT.toString())
+                                        && !clientService.existsByIdAndCpf(address.getClientId(), cpf))
+                                return ResponseEntity.notFound().build();
+
+                }
+
+                return ResponseEntity
+                                .ok(addressMapper.toDto(addressService.create(addressMapper.toEntity(address))));
         }
 
         @Operation(summary = "Get address by ID", description = "Retrieves a specific address by its ID")
@@ -62,9 +84,23 @@ public class AddressController {
         @GetMapping("/{id}")
         public ResponseEntity<AddressDTO> getAddressById(
                         @Parameter(description = "ID of the address to be retrieved") @PathVariable("id") Long id) {
-                Optional<Address> address = addressService.findById(id);
+                Optional<Address> address = null;
 
-                if (address.isPresent())
+                Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+                if (authentication instanceof CustomAuthentication) {
+                        CustomAuthentication customAuth = (CustomAuthentication) authentication;
+                        String role = customAuth.getRole();
+                        String cpf = customAuth.getCpf();
+
+                        if (role.equals(Role.CLIENT.toString()))
+                                address = addressService.getByIdAndUserCpf(id, cpf);
+                        if (role.equals(Role.ADMIN.toString()))
+                                address = addressService.findById(id);
+
+                }
+
+                if (address != null && address.isPresent())
                         return ResponseEntity.ok(addressMapper.toDto(address.get()));
 
                 return ResponseEntity.notFound().build();
@@ -81,7 +117,23 @@ public class AddressController {
         @RolesAllowed({ "ADMIN" })
         @GetMapping
         public Page<AddressDTO> getAllAddresses(@Parameter(hidden = true) Pageable pageable) {
-                return addressMapper.pageToPageDTO(addressService.findAll(pageable));
+                Page<Address> address = null;
+
+                Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+                if (authentication instanceof CustomAuthentication) {
+                        CustomAuthentication customAuth = (CustomAuthentication) authentication;
+                        String role = customAuth.getRole();
+                        String cpf = customAuth.getCpf();
+
+                        if (role.equals(Role.CLIENT.toString()))
+                                address = addressService.findAll(pageable);
+                        if (role.equals(Role.ADMIN.toString()))
+                                address = addressService.getAllByUserCpf(cpf, pageable);
+
+                }
+
+                return addressMapper.pageToPageDTO(address);
         }
 
         @Operation(summary = "Update an existing address", description = "Updates an existing address record")
@@ -96,8 +148,25 @@ public class AddressController {
         @PutMapping("/{id}")
         public AddressDTO updateAddress(
                         @Parameter(description = "ID of the address to be updated") @PathVariable("id") Long id,
-                        @RequestBody AddressDTO address) {
-                return addressMapper.toDto(addressService.update(id, addressMapper.toEntity(address)));
+                        @RequestBody AddressDTO addressDTO) {
+                Address address = null;
+
+                Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+                if (authentication instanceof CustomAuthentication) {
+                        CustomAuthentication customAuth = (CustomAuthentication) authentication;
+                        String role = customAuth.getRole();
+                        String cpf = customAuth.getCpf();
+
+                        if (role.equals(Role.CLIENT.toString()))
+                                address = addressService.updateByIdAndUserCpf(id, cpf,
+                                                addressMapper.toEntity(addressDTO));
+                        if (role.equals(Role.ADMIN.toString()))
+                                address = addressService.update(id, addressMapper.toEntity(addressDTO));
+
+                }
+
+                return addressMapper.toDto(address);
         }
 
         @Operation(summary = "Delete an address", description = "Deletes an address record by its ID")
@@ -111,6 +180,19 @@ public class AddressController {
         @DeleteMapping("/{id}")
         public boolean deleteAddress(
                         @Parameter(description = "ID of the address to be deleted") @PathVariable("id") Long id) {
-                return addressService.delete(id);
+                Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+                if (authentication instanceof CustomAuthentication) {
+                        CustomAuthentication customAuth = (CustomAuthentication) authentication;
+                        String role = customAuth.getRole();
+                        String cpf = customAuth.getCpf();
+
+                        if (role.equals(Role.CLIENT.toString()))
+                                return addressService.deleteByIdAndUserCpf(id, cpf);
+                        if (role.equals(Role.ADMIN.toString()))
+                                return addressService.delete(id);
+                }
+
+                return false;
         }
 }
