+++ src/main/java/com/pedro/petshop/controllers/BreedController.java
@@ -5,6 +5,8 @@ import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -14,11 +16,14 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.pedro.petshop.configs.CustomAuthentication;
 import com.pedro.petshop.configs.RolesAllowed;
 import com.pedro.petshop.dtos.BreedDTO;
 import com.pedro.petshop.entities.Breed;
+import com.pedro.petshop.enums.Role;
 import com.pedro.petshop.mappers.BreedMapper;
 import com.pedro.petshop.services.BreedService;
+import com.pedro.petshop.services.PetService;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -31,10 +36,12 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 public class BreedController {
 
         private final BreedService breedService;
+        private final PetService petService;
         private final BreedMapper breedMapper;
 
-        public BreedController(BreedService breedService, BreedMapper breedMapper) {
+        public BreedController(BreedService breedService, PetService petService, BreedMapper breedMapper) {
                 this.breedService = breedService;
+                this.petService = petService;
                 this.breedMapper = breedMapper;
         }
 
@@ -47,8 +54,21 @@ public class BreedController {
         })
         @RolesAllowed({ "ADMIN" })
         @PostMapping
-        public BreedDTO createBreed(@RequestBody BreedDTO breed) {
-                return breedMapper.toDto(breedService.create(breedMapper.toEntity(breed)));
+        public ResponseEntity<BreedDTO> createBreed(@RequestBody BreedDTO breed) {
+                Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+                if (authentication instanceof CustomAuthentication) {
+                        CustomAuthentication customAuth = (CustomAuthentication) authentication;
+                        String role = customAuth.getRole();
+                        String cpf = customAuth.getCpf();
+
+                        if (role.equals(Role.CLIENT.toString())
+                                        && !petService.existsByIdAndUserCpf(breed.getPetId(), cpf))
+                                return ResponseEntity.notFound().build();
+
+                }
+
+                return ResponseEntity.ok(breedMapper.toDto(breedService.create(breedMapper.toEntity(breed))));
         }
 
         @Operation(summary = "Get breed by ID", description = "Retrieves a specific breed by its ID")
@@ -62,9 +82,23 @@ public class BreedController {
         @GetMapping("/{id}")
         public ResponseEntity<BreedDTO> getBreedById(
                         @Parameter(description = "ID of the breed to be retrieved") @PathVariable("id") Long id) {
-                Optional<Breed> breed = breedService.findById(id);
+                Optional<Breed> breed = null;
 
-                if (breed.isPresent())
+                Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+                if (authentication instanceof CustomAuthentication) {
+                        CustomAuthentication customAuth = (CustomAuthentication) authentication;
+                        String role = customAuth.getRole();
+                        String cpf = customAuth.getCpf();
+
+                        if (role.equals(Role.CLIENT.toString()))
+                                breed = breedService.getByIdAndUserCpf(id, cpf);
+                        if (role.equals(Role.ADMIN.toString()))
+                                breed = breedService.findById(id);
+
+                }
+
+                if (breed != null && breed.isPresent())
                         return ResponseEntity.ok(breedMapper.toDto(breed.get()));
 
                 return ResponseEntity.notFound().build();
@@ -81,7 +115,23 @@ public class BreedController {
         @RolesAllowed({ "ADMIN" })
         @GetMapping
         public Page<BreedDTO> getAllBreeds(@Parameter(hidden = true) Pageable pageable) {
-                return breedMapper.pageToPageDTO(breedService.findAll(pageable));
+                Page<Breed> breeds = null;
+
+                Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+                if (authentication instanceof CustomAuthentication) {
+                        CustomAuthentication customAuth = (CustomAuthentication) authentication;
+                        String role = customAuth.getRole();
+                        String cpf = customAuth.getCpf();
+
+                        if (role.equals(Role.CLIENT.toString()))
+                                breeds = breedService.findAll(pageable);
+                        if (role.equals(Role.ADMIN.toString()))
+                                breeds = breedService.getAllByUserCpf(cpf, pageable);
+
+                }
+
+                return breedMapper.pageToPageDTO(breeds);
         }
 
         @Operation(summary = "Update an existing breed", description = "Updates an existing breed record")
@@ -96,8 +146,25 @@ public class BreedController {
         @PutMapping("/{id}")
         public BreedDTO updateBreed(
                         @Parameter(description = "ID of the breed to be updated") @PathVariable("id") Long id,
-                        @RequestBody BreedDTO breed) {
-                return breedMapper.toDto(breedService.update(id, breedMapper.toEntity(breed)));
+                        @RequestBody BreedDTO breedDTO) {
+                Breed breed = null;
+
+                Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+                if (authentication instanceof CustomAuthentication) {
+                        CustomAuthentication customAuth = (CustomAuthentication) authentication;
+                        String role = customAuth.getRole();
+                        String cpf = customAuth.getCpf();
+
+                        if (role.equals(Role.CLIENT.toString()))
+                                breed = breedService.updateByIdAndUserCpf(id, cpf,
+                                                breedMapper.toEntity(breedDTO));
+                        if (role.equals(Role.ADMIN.toString()))
+                                breed = breedService.update(id, breedMapper.toEntity(breedDTO));
+
+                }
+
+                return breedMapper.toDto(breed);
         }
 
         @Operation(summary = "Delete a breed", description = "Deletes a breed record by its ID")
@@ -111,6 +178,19 @@ public class BreedController {
         @DeleteMapping("/{id}")
         public boolean deleteBreed(
                         @Parameter(description = "ID of the breed to be deleted") @PathVariable("id") Long id) {
-                return breedService.delete(id);
+                Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+                if (authentication instanceof CustomAuthentication) {
+                        CustomAuthentication customAuth = (CustomAuthentication) authentication;
+                        String role = customAuth.getRole();
+                        String cpf = customAuth.getCpf();
+
+                        if (role.equals(Role.CLIENT.toString()))
+                                return breedService.deleteByIdAndUserCpf(id, cpf);
+                        if (role.equals(Role.ADMIN.toString()))
+                                return breedService.delete(id);
+                }
+
+                return false;
         }
 }
