+++ src/main/java/com/pedro/petshop/controllers/ClientController.java
@@ -9,6 +9,8 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -20,9 +22,11 @@ import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
+import com.pedro.petshop.configs.CustomAuthentication;
 import com.pedro.petshop.configs.RolesAllowed;
 import com.pedro.petshop.dtos.ClientDTO;
 import com.pedro.petshop.entities.Client;
+import com.pedro.petshop.enums.Role;
 import com.pedro.petshop.mappers.ClientMapper;
 import com.pedro.petshop.services.ClientService;
 
@@ -51,15 +55,26 @@ public class ClientController {
                         @ApiResponse(responseCode = "403", description = "Forbidden"),
                         @ApiResponse(responseCode = "500", description = "Internal server error")
         })
-        @RolesAllowed({ "ADMIN" })
+        @RolesAllowed({ "ADMIN", "CLIENT" })
         @PostMapping(value = "/{id}/upload-image", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
         public HttpStatus uploadProfileImage(@PathVariable("id") Long id,
                         @RequestPart(value = "file", required = true) MultipartFile file) {
+                Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+                if (authentication instanceof CustomAuthentication) {
+                        CustomAuthentication customAuth = (CustomAuthentication) authentication;
+                        String role = customAuth.getRole();
+                        String cpf = customAuth.getCpf();
+
+                        if (role.equals(Role.CLIENT.toString()) && !clientService.existsByIdAndCpf(id, cpf))
+                                return HttpStatus.NOT_FOUND;
+                }
+
                 boolean isSaved = clientService.uploadImage(id, file);
                 if (isSaved)
                         return HttpStatus.OK;
 
-                return HttpStatus.INTERNAL_SERVER_ERROR;
+                return HttpStatus.NOT_FOUND;
         }
 
         @Operation(summary = "Download a client image", description = "Download a client image from the system")
@@ -69,9 +84,20 @@ public class ClientController {
                         @ApiResponse(responseCode = "403", description = "Forbidden"),
                         @ApiResponse(responseCode = "404", description = "Not found"),
         })
-        @RolesAllowed({ "ADMIN" })
+        @RolesAllowed({ "ADMIN", "CLIENT" })
         @GetMapping("/{id}/download-image")
         public ResponseEntity<Resource> getProfileImage(@PathVariable("id") Long id) {
+                Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+                if (authentication instanceof CustomAuthentication) {
+                        CustomAuthentication customAuth = (CustomAuthentication) authentication;
+                        String role = customAuth.getRole();
+                        String cpf = customAuth.getCpf();
+
+                        if (role.equals(Role.CLIENT.toString()) && !clientService.existsByIdAndCpf(id, cpf))
+                                return ResponseEntity.notFound().build();
+                }
+
                 Optional<Resource> resource = clientService.getProfileImage(id);
                 if (!resource.isPresent())
                         return ResponseEntity.notFound().build();
@@ -90,10 +116,23 @@ public class ClientController {
                         @ApiResponse(responseCode = "403", description = "Forbidden"),
                         @ApiResponse(responseCode = "400", description = "Invalid input data")
         })
-        @RolesAllowed({ "ADMIN" })
+        @RolesAllowed({ "ADMIN", "CLIENT" })
         @PostMapping
-        public ClientDTO createClient(@RequestBody ClientDTO client) {
-                return clientMapper.toDto(clientService.create(clientMapper.toEntity(client)));
+        public ResponseEntity<ClientDTO> createClient(@RequestBody ClientDTO client) {
+                Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+                if (authentication instanceof CustomAuthentication) {
+                        CustomAuthentication customAuth = (CustomAuthentication) authentication;
+                        String role = customAuth.getRole();
+                        String cpf = customAuth.getCpf();
+
+                        if (role.equals(Role.CLIENT.toString())) {
+                                client.setCpf(cpf);
+                        }
+                }
+
+                return ResponseEntity
+                                .ok(clientMapper.toDto(clientService.create(clientMapper.toEntity(client))));
         }
 
         @Operation(summary = "Get client by ID", description = "Retrieves a specific client by its ID")
@@ -103,12 +142,26 @@ public class ClientController {
                         @ApiResponse(responseCode = "403", description = "Forbidden"),
                         @ApiResponse(responseCode = "404", description = "Client not found")
         })
-        @RolesAllowed({ "ADMIN" })
+        @RolesAllowed({ "ADMIN", "CLIENT" })
         @GetMapping("/{id}")
         public ResponseEntity<ClientDTO> getClientById(@PathVariable("id") Long id) {
-                Optional<Client> client = clientService.findById(id);
+                Optional<Client> client = null;
 
-                if (client.isPresent())
+                Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+                if (authentication instanceof CustomAuthentication) {
+                        CustomAuthentication customAuth = (CustomAuthentication) authentication;
+                        String role = customAuth.getRole();
+                        String cpf = customAuth.getCpf();
+
+                        if (role.equals(Role.CLIENT.toString()))
+                                client = clientService.getByIdAndCpf(id, cpf);
+                        if (role.equals(Role.ADMIN.toString()))
+                                client = clientService.findById(id);
+
+                }
+
+                if (client != null && client.isPresent())
                         return ResponseEntity.ok(clientMapper.toDto(client.get()));
 
                 return ResponseEntity.notFound().build();
@@ -122,10 +175,26 @@ public class ClientController {
                         @ApiResponse(responseCode = "401", description = "Unauthorized"),
                         @ApiResponse(responseCode = "403", description = "Forbidden"),
         })
-        @RolesAllowed({ "ADMIN" })
+        @RolesAllowed({ "ADMIN", "CLIENT" })
         @GetMapping
         public Page<ClientDTO> getAllClients(@Parameter(hidden = true) Pageable pageable) {
-                return clientMapper.pageToPageDTO(clientService.findAll(pageable));
+                Page<Client> clients = null;
+
+                Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+                if (authentication instanceof CustomAuthentication) {
+                        CustomAuthentication customAuth = (CustomAuthentication) authentication;
+                        String role = customAuth.getRole();
+                        String cpf = customAuth.getCpf();
+
+                        if (role.equals(Role.CLIENT.toString()))
+                                clients = clientService.findAll(pageable);
+                        if (role.equals(Role.ADMIN.toString()))
+                                clients = clientService.getAllByCpf(cpf, pageable);
+
+                }
+
+                return clientMapper.pageToPageDTO(clients);
         }
 
         @Operation(summary = "Update an existing client", description = "Updates an existing client record")
@@ -136,12 +205,28 @@ public class ClientController {
                         @ApiResponse(responseCode = "403", description = "Forbidden"),
                         @ApiResponse(responseCode = "404", description = "Client not found")
         })
-        @RolesAllowed({ "ADMIN" })
+        @RolesAllowed({ "ADMIN", "CLIENT" })
         @PutMapping("/{id}")
         public ClientDTO updateClient(
                         @Parameter(description = "ID of the client to be updated") @PathVariable("id") Long id,
-                        @RequestBody ClientDTO client) {
-                return clientMapper.toDto(clientService.update(id, clientMapper.toEntity(client)));
+                        @RequestBody ClientDTO clientDTO) {
+                Client client = null;
+
+                Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+                if (authentication instanceof CustomAuthentication) {
+                        CustomAuthentication customAuth = (CustomAuthentication) authentication;
+                        String role = customAuth.getRole();
+                        String cpf = customAuth.getCpf();
+
+                        if (role.equals(Role.CLIENT.toString()))
+                                client = clientService.updateByIdAndCpf(id, cpf, clientMapper.toEntity(clientDTO));
+                        if (role.equals(Role.ADMIN.toString()))
+                                client = clientService.update(id, clientMapper.toEntity(clientDTO));
+
+                }
+
+                return clientMapper.toDto(client);
         }
 
         @Operation(summary = "Delete a client", description = "Deletes a client record by its ID")
@@ -151,10 +236,23 @@ public class ClientController {
                         @ApiResponse(responseCode = "403", description = "Forbidden"),
                         @ApiResponse(responseCode = "404", description = "Client not found")
         })
-        @RolesAllowed({ "ADMIN" })
+        @RolesAllowed({ "ADMIN", "CLIENT" })
         @DeleteMapping("/{id}")
         public boolean deleteClient(
                         @Parameter(description = "ID of the client to be deleted") @PathVariable("id") Long id) {
-                return clientService.delete(id);
+                Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+                if (authentication instanceof CustomAuthentication) {
+                        CustomAuthentication customAuth = (CustomAuthentication) authentication;
+                        String role = customAuth.getRole();
+                        String cpf = customAuth.getCpf();
+
+                        if (role.equals(Role.CLIENT.toString()))
+                                return clientService.deleteByIdAndCpf(id, cpf);
+                        if (role.equals(Role.ADMIN.toString()))
+                                return clientService.delete(id);
+                }
+
+                return false;
         }
 }
